requirements.txt
aws-psycopg2
peewee
graphene
pyjwt[crypto]==2.3.0
git+https://gitlab.com/gabano2005/vault-client.git@master#egg=vault-client&subdirectory=src
